pyproject.toml
[project]
name = "insurance-be"
version = "0.1.0"
description = "A Django REST API project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular[sidecar]>=0.28.0",
    
]
